test: move the oversized-header-list subprocess tests to h2-conformance.test.ts
node-http2.test.js is already at its ASAN timeout budget with the 10k-request
maxSessionMemory stress test; the subprocess tests belong with the other
wire-level conformance cases anyway. Give the stress test a bit more headroom
(it was already at ~150s in capped-core environments on main).
